safari-600.7-branch/LayoutTests/fast/dom/shadow/style-scoped-not-enabled-expected.txt

This test ensures content element feature can be disabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(shouldBeApplied).color is 'rgb(0, 0, 255)'
PASS window.getComputedStyle(shouldNotBeApplied).color is 'rgb(0, 0, 0)'
PASS window.getComputedStyle(insideScope).color is 'rgb(255, 0, 0)'
PASS window.getComputedStyle(shouldNotBeScoped).color is 'rgb(255, 0, 0)'
PASS successfullyParsed is true

TEST COMPLETE
Should not be blue
Should be red

Should be red
